% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_to_num.R
\name{day_to_num}
\alias{day_to_num}
\title{Convert the days of the week to numeric}
\usage{
day_to_num(x)
}
\arguments{
\item{x}{A character string containing days of the week in full (e.g.
"Friday") or abbreviated (e.g. "Fri") form.}
}
\description{
Convert the days of the week to numeric
}
