% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/del_dir.R
\name{del_dir}
\alias{del_dir}
\title{Delete files from one directory which don't exist in another}
\usage{
del_dir(dir_one, dir_two)
}
\arguments{
\item{dir_one}{The first directory.}

\item{dir_two}{The second directory.}
}
\description{
Delete files from one directory which don't exist in another
}
