% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrow_path.R
\name{nrow_path}
\alias{nrow_path}
\title{Calculate the number of rows in an unloaded file}
\usage{
nrow_path(filepath = getwd(), pattern = NULL, col_names = TRUE)
}
\arguments{
\item{filepath}{The filepath.}

\item{pattern}{Regular expression.}

\item{col_names}{Are column names present?}
}
\description{
Calculate the number of rows in an unloaded file
}
