% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_missing.R
\name{prop_missing}
\alias{prop_missing}
\title{Calculate the proportion of missing values in each column}
\usage{
prop_missing(df)
}
\arguments{
\item{df}{The dataframe.}
}
\description{
Calculate the proportion of missing values in each column
}
