% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_scripts.R
\name{source_scripts}
\alias{source_scripts}
\title{Source all R scripts in a directory}
\usage{
source_scripts(filepath = getwd())
}
\arguments{
\item{filepath}{A character string denoting a filepath. Defaults to the
working directory, \code{getwd()}. \code{source_scripts} errors if supplied
with an invalid directory name or with a directory containing no R scripts.}
}
\description{
\code{source_scripts} sources all R scripts within a given
directory in a single function call.
}
\examples{
\dontrun{
source_scripts()
}

}
\seealso{
\code{\link[base]{source}} for sourcing a single R script.
}
