% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_to_num.R
\name{day_to_num}
\alias{day_to_num}
\title{Convert the days of the week to numeric}
\usage{
day_to_num(x)
}
\arguments{
\item{x}{A character string containing days of the week in full (e.g.
"Friday") or abbreviated (e.g. "Fri") form.}
}
\description{
\code{day_to_num} converts the days of the week from
\code{character} format into numeric format for use in \code{\link{n_day}}.
}
\details{
The \code{Date} and \code{POSIXt} classes within R treat Sunday as
day 0 in a week, and Saturday as day 6. \code{day_to_num} also treats Sunday
as day 0 (as opposed to day 7) for the sake of consistency.
}
\examples{
day_to_num(c("Monday", "Tuesday", "Wed"))
day_to_num(days_of_week())
day_to_num(weekdays(lubridate::dmy(01012020)))

}
\seealso{
\code{\link{days_of_week}} for returning a vector of the days of the week.

\code{\link[base]{weekdays}} for converting a date into the day of the week.

\code{\link{n_day}} for calculating the number of occurrences of a day of
the week in a set time period.
}
