% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_of_week.R
\name{days_of_week}
\alias{days_of_week}
\title{The days of the week}
\usage{
days_of_week(abbr = FALSE)
}
\arguments{
\item{abbr}{Logical. If FALSE, returns the days of the week in full form
(e.g. "Friday"). If TRUE, it returns days in abbreviated form (e.g. "Fri").
The default value is FALSE.}
}
\description{
The days of the week
}
