% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jafun.R
\docType{package}
\name{jafun}
\alias{jafun}
\title{\code{jafun} package}
\description{
Personal R package of Jack Hannah.
}
\details{
See the README on
\href{https://github.com/jackhannah95/jafun#readme}{GitHub}.
}
